Exclude items from the Property Model creation
Prevent has_audio and has_video property edit from UI when
corresponding Track was set to disable audio or video.

Turns out that removing property form the model is the fastest and only
effective way to temporary disable any kind of edits that are available
for the user over the items.